Image update problem solve

+++ app/Controllers/Admin/Awards.php
@@ -152,7 +152,10 @@ class Awards extends BaseController
             //old image unlink
             $old_img = get_data_by_id('value','global_settings','id',$id);
             if (!empty($old_img)){
-                unlink($target_dir.''.$old_img);
+                $imgPath = $target_dir.''.$old_img;
+                if (file_exists($imgPath)) {
+                    unlink($target_dir . '' . $old_img);
+                }
             }
 
             //new image uplode
