+++ app/Controllers/Admin/Podcasts.php
@@ -137,7 +137,10 @@ class Podcasts extends BaseController
             //old image unlink
             $old_img = get_data_by_id('value','global_settings','id',$id);
             if (!empty($old_img)){
-                unlink($target_dir.''.$old_img);
+                $imgPath = $target_dir.''.$old_img;
+                if (file_exists($imgPath)) {
+                    unlink($target_dir . '' . $old_img);
+                }
             }
 
             //new image uplode
