+++ app/Controllers/Admin/Publication.php
@@ -207,7 +207,10 @@ class Publication extends BaseController
             //old image unlink
             $old_img = get_data_by_id('value','global_settings','id',$id);
             if (!empty($old_img)){
-                unlink($target_dir.''.$old_img);
+                $imgPath = $target_dir.''.$old_img;
+                if (file_exists($imgPath)) {
+                    unlink($target_dir . '' . $old_img);
+                }
             }
 
             //new image uplode
@@ -237,7 +240,10 @@ class Publication extends BaseController
         $target_dir = FCPATH . '/uploads/publication/';
         $old_img = get_data_by_id('image','publication','publication_id',$id);
         if (!empty($old_img)){
-            unlink($target_dir.''.$old_img);
+            $imgPath = $target_dir.''.$old_img;
+            if (file_exists($imgPath)) {
+                unlink($target_dir . '' . $old_img);
+            }
         }
 
         $table = DB()->table('publication');
