+++ app/Controllers/Admin/Settings.php
@@ -124,7 +124,10 @@ class Settings extends BaseController
             //old image unlink
             $old_img = get_data_by_id('pic','admin','user_id',$data['user_id']);
             if (!empty($old_img)){
-                unlink($target_dir.''.$old_img);
+                $imgPath = $target_dir.''.$old_img;
+                if (file_exists($imgPath)) {
+                    unlink($target_dir . '' . $old_img);
+                }
             }
 
             //new image uplode
