+++ app/Controllers/Admin/Slider.php
@@ -12,24 +12,25 @@ class Slider extends BaseController
 
     public function __construct()
     {
-        $this->validation =  \Config\Services::validation();
+        $this->validation = \Config\Services::validation();
         $this->session = \Config\Services::session();
         $this->crop = \Config\Services::image();
     }
+
     public function index()
     {
         $isLoggedIn = $this->session->isLoggedIn;
         $role_id = $this->session->role;
         if (!isset($isLoggedIn) || $isLoggedIn != TRUE) {
             return redirect()->to(site_url('Admin/login'));
-        }else {
+        } else {
             $table = DB()->table('slider');
             $data['slider'] = $table->get()->getResult();
 
 
             echo view('Admin/header');
             echo view('Admin/sidebar');
-            echo view('Admin/Slider/list',$data);
+            echo view('Admin/Slider/list', $data);
             echo view('Admin/footer');
         }
     }
@@ -40,31 +41,32 @@ class Slider extends BaseController
         $role_id = $this->session->role;
         if (!isset($isLoggedIn) || $isLoggedIn != TRUE) {
             return redirect()->to(site_url('Admin/login'));
-        }else {
+        } else {
 
             $data['action'] = base_url('Admin/Slider/create_action');
 
 
             echo view('Admin/header');
             echo view('Admin/sidebar');
-            echo view('Admin/Slider/create',$data);
+            echo view('Admin/Slider/create', $data);
             echo view('Admin/footer');
         }
     }
 
-    public function create_action(){
+    public function create_action()
+    {
 
         $data['name'] = $this->request->getPost('name');
-        if (!empty($_FILES['image']['name'])){
+        if (!empty($_FILES['image']['name'])) {
             $target_dir = FCPATH . '/uploads/slider_img/';
-            if(!file_exists($target_dir)){
-                mkdir($target_dir,0777);
+            if (!file_exists($target_dir)) {
+                mkdir($target_dir, 0777);
             }
 
             //new image uplode
             $pic = $this->request->getFile('image');
-            $namePic = 'slider_'.$pic->getRandomName();//$pic->getRandomName();
-            $pic->move($target_dir,$namePic);
+            $namePic = 'slider_' . $pic->getRandomName();//$pic->getRandomName();
+            $pic->move($target_dir, $namePic);
 //            $news_img = 'news_'.$pic->getName();
 //            $this->crop->withFile($target_dir.''.$namePic)->fit(1000, 1000, 'center')->save($target_dir.''.$news_img);
 //            unlink($target_dir.''.$namePic);
@@ -78,7 +80,7 @@ class Slider extends BaseController
         if ($this->validation->run($data) == FALSE) {
             $this->session->setFlashdata('message', '<div class="alert alert-danger alert-dismissible" role="alert">' . $this->validation->listErrors() . ' <button type="button" class="close" data-dismiss="alert" aria-label="Close"><span aria-hidden="true">&times;</span></button></div>');
             return redirect()->to('/Admin/Slider/create');
-        }else{
+        } else {
             $table = DB()->table('slider');
             $table->insert($data);
 
@@ -88,46 +90,51 @@ class Slider extends BaseController
     }
 
 
-    public function update($id){
+    public function update($id)
+    {
         $isLoggedIn = $this->session->isLoggedIn;
         $role_id = $this->session->role;
         if (!isset($isLoggedIn) || $isLoggedIn != TRUE) {
             return redirect()->to(site_url('Admin/login'));
-        }else {
+        } else {
             $table = DB()->table('slider');
-            $data['slider'] = $table->where('sl_id',$id)->get()->getRow();
+            $data['slider'] = $table->where('sl_id', $id)->get()->getRow();
             $data['action'] = base_url('Admin/Slider/update_action');
 
             echo view('Admin/header');
             echo view('Admin/sidebar');
-            echo view('Admin/Slider/update',$data);
+            echo view('Admin/Slider/update', $data);
             echo view('Admin/footer');
         }
     }
 
-    public function update_action(){
+    public function update_action()
+    {
 
         $data['sl_id'] = $this->request->getPost('sl_id');
 
         $data['name'] = $this->request->getPost('name');
         $data['status'] = $this->request->getPost('status');
 
-        if (!empty($_FILES['image']['name'])){
+        if (!empty($_FILES['image']['name'])) {
             $target_dir = FCPATH . '/uploads/slider_img/';
-            if(!file_exists($target_dir)){
-                mkdir($target_dir,0777);
+            if (!file_exists($target_dir)) {
+                mkdir($target_dir, 0777);
             }
 
             //old image unlink
-            $old_img = get_data_by_id('image','slider','sl_id',$data['sl_id']);
-            if (!empty($old_img)){
-                unlink($target_dir.''.$old_img);
+            $old_img = get_data_by_id('image', 'slider', 'sl_id', $data['sl_id']);
+            if (!empty($old_img)) {
+                $imgPath = $target_dir.''.$old_img;
+                if (file_exists($imgPath)) {
+                    unlink($target_dir . '' . $old_img);
+                }
             }
 
             //new image uplode
             $pic = $this->request->getFile('image');
-            $namePic = 'news_'.$pic->getRandomName();//$pic->getRandomName();
-            $pic->move($target_dir,$namePic);
+            $namePic = 'news_' . $pic->getRandomName();//$pic->getRandomName();
+            $pic->move($target_dir, $namePic);
 //            $news_img = 'news_'.$pic->getName();
 //            $this->crop->withFile($target_dir.''.$namePic)->fit(1000, 1000, 'center')->save($target_dir.''.$news_img);
 //            unlink($target_dir.''.$namePic);
@@ -142,34 +149,35 @@ class Slider extends BaseController
 
         if ($this->validation->run($data) == FALSE) {
             $this->session->setFlashdata('message', '<div class="alert alert-danger alert-dismissible" role="alert">' . $this->validation->listErrors() . ' <button type="button" class="close" data-dismiss="alert" aria-label="Close"><span aria-hidden="true">&times;</span></button></div>');
-            return redirect()->to('/Admin/Slider/update/'.$data['sl_id']);
-        }else{
+            return redirect()->to('/Admin/Slider/update/' . $data['sl_id']);
+        } else {
             $table = DB()->table('slider');
-            $table->where('sl_id',$data['sl_id'])->update($data);
+            $table->where('sl_id', $data['sl_id'])->update($data);
 
             $this->session->setFlashdata('message', '<div class="alert alert-success alert-dismissible" role="alert">Update Record Success <button type="button" class="close" data-dismiss="alert" aria-label="Close"><span aria-hidden="true">&times;</span></button></div>');
-            return redirect()->to('/Admin/Slider/update/'.$data['sl_id']);
+            return redirect()->to('/Admin/Slider/update/' . $data['sl_id']);
         }
     }
 
-    public function delete($id){
+    public function delete($id)
+    {
 
         //old image unlink
-        $old_img = get_data_by_id('image','slider','sl_id',$id);
+        $old_img = get_data_by_id('image', 'slider', 'sl_id', $id);
         $target_dir = FCPATH . '/uploads/slider_img/';
-        if (!empty($old_img)){
-            unlink($target_dir.''.$old_img);
+        if (!empty($old_img)) {
+            $imgPath = $target_dir.''.$old_img;
+            if (file_exists($imgPath)) {
+                unlink($target_dir . '' . $old_img);
+            }
         }
 
         $table = DB()->table('slider');
-        $table->where('sl_id',$id)->delete();
+        $table->where('sl_id', $id)->delete();
 
         $this->session->setFlashdata('message', '<div class="alert alert-success alert-dismissible" role="alert">Delete Record Success <button type="button" class="close" data-dismiss="alert" aria-label="Close"><span aria-hidden="true">&times;</span></button></div>');
         return redirect()->to('/Admin/Slider/');
     }
-
-
-
 
 
 }
